refactor: isRegistrationPeriod 파일 위치를 features에서 entities로 이동
