+++ src/entities/club/config.ts
